07: Rename components and reorganise export
