+++ homeworks/07-mars-viewer/src/App.module.css
@@ -17,6 +17,6 @@
 }
 
 .title {
-  font-size: 4.4rem;
+  font-size: 3.3rem;
   font-family: "Gill Sans", "Gill Sans MT", Calibri, "Trebuchet MS", sans-serif;
 }
