+++ homeworks/07-mars-viewer/src/App.tsx
@@ -1,15 +1,14 @@
 import React from 'react';
 import styles from './App.module.css';
 
-import { SolSelectForm } from './components/SolSelectForm';
-import { SolPhotos } from './components/SolPhotos';
+import { PhotoLoader, PhotoViewer } from './components';
 
 export const App: React.FC = () => (
   <div className="App">
     <div className={styles.container}>
       <h1 className={styles.title}>Select Sol and press "load"!</h1>
-      <SolSelectForm />
-      <SolPhotos />
+      <PhotoLoader />
+      <PhotoViewer />
     </div>
   </div>
 );
