+++ homeworks/07-mars-viewer/src/components/PhotoLoader.module.css
