+++ homeworks/07-mars-viewer/src/components/index.ts
@@ -1,0 +1,3 @@
+export { PhotoCard } from './PhotoCard';
+export { PhotoLoader } from './PhotoLoader';
+export { PhotoViewer } from './PhotoViewer';
